+++ runtime/docs/CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Rename `app_staking_adjustment` to `app_session_tokens_multiplier`
 - Remove `app_baseline_stake_rate`
+- Rename `keygenerator` to `keygen`
 
 ## [0.0.0.23] - 2023-02-21